Time: 2 ms (100%), Space: 94.9 MB (80.32%) - LeetHub

+++ 2640-find-the-score-of-all-prefixes-of-an-array/2640-find-the-score-of-all-prefixes-of-an-array.java
@@ -2,10 +2,10 @@ class Solution {
     public long[] findPrefixScore(int[] nums) {
         int n = nums.length;
         long[] c = new long[n];
-        int max = nums[0];
+        long max = nums[0];
         for(int i = 0; i<n; i++){
             if(nums[i] > max) max = nums[i];
-            c[i] = (long) (nums[i] + max);
+            c[i] = nums[i] + max;
         }
         for(int i = 1; i<n; i++){
             c[i] += c[i-1];
